+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/additionalconfig/DefaultAdditionalUIConfigProvider.scala
@@ -3,11 +3,11 @@ package pl.touk.nussknacker.ui.additionalconfig
 import pl.touk.nussknacker.engine.api.component.{ComponentId, ScenarioPropertyConfig}
 
 class DefaultAdditionalUIConfigProvider(
-    processingTypeToConfig: Map[String, Map[ComponentId, AdditionalUIConfig]],
+    processingTypeToConfig: Map[String, Map[ComponentId, ComponentAdditionalConfig]],
     processingTypeToAdditionalPropertiesConfig: Map[String, Map[String, ScenarioPropertyConfig]],
 ) extends AdditionalUIConfigProvider {
 
-  override def getAllForProcessingType(processingType: String): Map[ComponentId, AdditionalUIConfig] =
+  override def getAllForProcessingType(processingType: String): Map[ComponentId, ComponentAdditionalConfig] =
     processingTypeToConfig.getOrElse(processingType, Map.empty)
 
   override def getScenarioPropertiesUIConfigs(processingType: String): Map[String, ScenarioPropertyConfig] =