+++ docs/Changelog.md
@@ -4,7 +4,7 @@
 * [#4287](https://github.com/TouK/nussknacker/pull/4287) Improvement: bump cats effect to 3.5.2
 * [#5257](https://github.com/TouK/nussknacker/pull/5257) Updated Flink 1.16.2 -> 1.16.3
 * [#5253](https://github.com/TouK/nussknacker/pull/5253) Removed the option to configure fragments via config. Due to the recent expansion of FragmentParameter, the option has become largely redundant. Removed to decrease unnecessary complexity.
-* [#5271](https://github.com/TouK/nussknacker/pull/5271) Changed AdditionalUIConfigProvider to be more in line with FragmentParameter (TODO more explanation?)
+* [#5271](https://github.com/TouK/nussknacker/pull/5271) Changed `AdditionalUIConfigProvider.getAllForProcessingType` API to be more in line with FragmentParameter
 
 1.13.0 (Not released yet)
 -------------------------